Refactorizar backend
