--- src/main/java/com/acme/nutrimove/platform/activities/domain/model/commands/DeleteActivityCommand.java
@@ -1,7 +1,0 @@
-package com.acme.nutrimove.platform.activities.domain.model.commands;
-
-public record DeleteActivityCommand(Long activityId) {
-    public DeleteActivityCommand {
-        if (activityId == null) throw new IllegalArgumentException("activityId is null");
-    }
-}
