--- src/main/java/com/acme/nutrimove/platform/activities/domain/model/queries/GetActivityByIdQuery.java
@@ -1,7 +1,0 @@
-package com.acme.nutrimove.platform.activities.domain.model.queries;
-
-public record GetActivityByIdQuery(Long id) {
-    public GetActivityByIdQuery {
-        if (id == null) throw new NullPointerException("id is null");
-    }
-}
