--- src/main/java/com/acme/nutrimove/platform/activities/domain/services/ActivityQueryService.java
@@ -1,16 +1,0 @@
-package com.acme.nutrimove.platform.activities.domain.services;
-
-import com.acme.nutrimove.platform.activities.domain.model.aggregates.Activity;
-import com.acme.nutrimove.platform.activities.domain.model.queries.GetActivityByIdQuery;
-import com.acme.nutrimove.platform.activities.domain.model.queries.GetAllActivityByNameQuery;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ActivityQueryService {
-    List<Activity> handle(GetAllActivityByNameQuery query);
-    List<Activity> getAllActivities();
-
-    Optional<Activity> handle(GetActivityByIdQuery query);
-
-}
