--- src/main/java/com/acme/nutrimove/platform/activities/interfaces/rest/resources/ActivityResource.java
@@ -1,4 +1,0 @@
-package com.acme.nutrimove.platform.activities.interfaces.rest.resources;
-
-public record ActivityResource(Long id, String name, String description, Integer duration, Long userId ) {
-}
