+++ src/main/java/com/acme/nutrimove/platform/backend/activities/domain/model/commands/DeleteActivityCommand.java
@@ -1,0 +1,7 @@
+package com.acme.nutrimove.platform.backend.activities.domain.model.commands;
+
+public record DeleteActivityCommand(Long activityId) {
+    public DeleteActivityCommand {
+        if (activityId == null) throw new IllegalArgumentException("activityId is null");
+    }
+}
