+++ src/main/java/com/acme/nutrimove/platform/backend/activities/domain/model/queries/GetActivityByIdQuery.java
@@ -1,0 +1,7 @@
+package com.acme.nutrimove.platform.backend.activities.domain.model.queries;
+
+public record GetActivityByIdQuery(Long id) {
+    public GetActivityByIdQuery {
+        if (id == null) throw new NullPointerException("id is null");
+    }
+}
