+++ src/main/java/com/acme/nutrimove/platform/backend/activities/domain/model/queries/GetAllActivityByNameQuery.java
@@ -1,0 +1,7 @@
+package com.acme.nutrimove.platform.backend.activities.domain.model.queries;
+
+public record GetAllActivityByNameQuery(String name) {
+    public GetAllActivityByNameQuery {
+        if (name == null) throw new NullPointerException("name is null");
+    }
+}
