+++ src/main/java/com/acme/nutrimove/platform/backend/activities/interfaces/rest/resources/ActivityResource.java
@@ -1,0 +1,4 @@
+package com.acme.nutrimove.platform.backend.activities.interfaces.rest.resources;
+
+public record ActivityResource(Long id, String name, String description, Integer duration, Long userId ) {
+}
