+++ src/main/java/com/acme/nutrimove/platform/backend/activities/interfaces/rest/transform/CreateActivityCommandFromResourceAssembler.java
@@ -1,0 +1,11 @@
+package com.acme.nutrimove.platform.backend.activities.interfaces.rest.transform;
+
+import com.acme.nutrimove.platform.backend.activities.domain.model.commands.CreateActivityCommand;
+import com.acme.nutrimove.platform.backend.activities.interfaces.rest.resources.CreateActivityResource;
+
+public class CreateActivityCommandFromResourceAssembler {
+
+    public static CreateActivityCommand toCommand(CreateActivityResource resource) {
+        return new CreateActivityCommand(resource.name(), resource.description(), resource.duration(), resource.userId());
+    }
+}
