+++ src/main/java/com/acme/nutrimove/platform/backend/subscriptions/domain/model/queries/GetSubscriptionByIdQuery.java
@@ -1,0 +1,7 @@
+package com.acme.nutrimove.platform.backend.subscriptions.domain.model.queries;
+
+public record GetSubscriptionByIdQuery(Long id) {
+    public GetSubscriptionByIdQuery {
+        if (id == null) throw new NullPointerException("id is null");
+    }
+}
