--- src/main/java/com/acme/nutrimove/platform/subscriptions/domain/model/queries/GetAllSubscriptionByDescriptionQuery.java
@@ -1,9 +1,0 @@
-package com.acme.nutrimove.platform.subscriptions.domain.model.queries;
-
-public record GetAllSubscriptionByDescriptionQuery(String description) {
-    public GetAllSubscriptionByDescriptionQuery {
-        if (description == null || description.isBlank()) {
-            throw new IllegalArgumentException("Description cannot be null or empty");
-        }
-    }
-}
