--- src/main/java/com/acme/nutrimove/platform/subscriptions/interfaces/rest/resources/SubscriptionResource.java
@@ -1,4 +1,0 @@
-package com.acme.nutrimove.platform.subscriptions.interfaces.rest.resources;
-
-public record SubscriptionResource(Long id, String description, Double price, Integer monthDuration, Boolean trial, Long userId) {
-}
